Renaming and rebuilding contexts
