--- src/OutboxPatternDemo.Publisher/CustomOutbox/CustomOutboxContext.cs
@@ -1,16 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-
-namespace OutboxPatternDemo.Publisher.CustomOutbox;
-
-public class CustomOutboxContext : DbContext
-{
-    public CustomOutboxContext(DbContextOptions<CustomOutboxContext> options)
-        : base(options)
-    {
-    }
-
-    internal DbSet<CustomOutboxMessage> Messages { get; set; }
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder) =>
-        modelBuilder.HasDefaultSchema("CustomOutbox");
-}
