--- src/OutboxPatternDemo.Publisher/CustomOutbox/CustomOutboxContextFactory.cs
@@ -1,16 +1,0 @@
-using System.Data.SqlClient;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Design;
-
-namespace OutboxPatternDemo.Publisher.CustomOutbox;
-
-public class CustomOutboxContextFactory : IDesignTimeDbContextFactory<CustomOutboxContext>
-{
-    public CustomOutboxContext CreateDbContext(string[] args)
-    {
-        var sqlConnection = new SqlConnection("Data Source=localhost;Initial Catalog=OutboxPatternDemo;Integrated Security=SSPI;TrustServerCertificate=True");
-        var contextOptionsBuilder = new DbContextOptionsBuilder<CustomOutboxContext>();
-        contextOptionsBuilder.UseSqlServer(sqlConnection);
-        return new CustomOutboxContext(contextOptionsBuilder.Options);
-    }
-}
