--- src/OutboxPatternDemo.Publisher/CustomOutbox/CustomOutboxMessage.cs
@@ -1,14 +1,0 @@
-using System;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace OutboxPatternDemo.Publisher.CustomOutbox;
-
-internal class CustomOutboxMessage
-{
-    [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-    public int Id { get; private set; }
-    public string Type { get; internal set; }
-    public string Data { get; internal set; }
-    public DateTime RequestedTimeUtc { get; internal set; }
-    public DateTime? ProcessedTimeUtc { get; internal set; }
-}
