+++ src/OutboxPatternDemo.Publisher/MedicalRecords/Data/MedicalRecordContext.cs
@@ -1,0 +1,16 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace OutboxPatternDemo.Publisher.MedicalRecords.Data;
+
+public class MedicalRecordContext : DbContext
+{
+    public MedicalRecordContext(DbContextOptions<MedicalRecordContext> options)
+        : base(options)
+    {
+    }
+
+    public DbSet<AppointmentNotesDto> AppointmentNotes { get; set; }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder) =>
+        modelBuilder.HasDefaultSchema("MedicalRecord");
+}
