+++ src/OutboxPatternDemo.Subscriber/DuplicateCheckers/Sql/Data/DuplicateKey.cs
@@ -1,0 +1,12 @@
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace OutboxPatternDemo.Subscriber.DuplicateCheckers.Sql.Data;
+
+public class DuplicateKey
+{
+    public DuplicateKey(string key) => Key = key;
+
+    [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+    public int Id { get; private set; }
+    public string Key { get; }
+}
